+++ server/distributed/server.go
@@ -2,13 +2,11 @@ package distributed
 
 import (
 	"context"
-	"strings"
 	"time"
 
 	"github.com/hyperledger/fabric-sdk-go/pkg/common/errors/status"
 	"github.com/hyperledger/fabric-sdk-go/pkg/core/config"
 	"github.com/owkin/orchestrator/lib/asset"
-	orcerrors "github.com/owkin/orchestrator/lib/errors"
 	"github.com/owkin/orchestrator/server/common"
 	"github.com/owkin/orchestrator/server/common/logger"
 	"github.com/owkin/orchestrator/server/common/trace"
@@ -74,18 +72,8 @@ func (a *AppServer) Stop() {
 // shouldRetry will trigger a retry on specific orchestration error.
 func shouldRetry(err error) bool {
 	st, ok := status.FromError(err)
-	msg := err.Error()
 	switch {
 	case ok && st.Code == int32(status.Timeout):
-		return true
-	case strings.Contains(msg, orcerrors.ErrReferenceNotFound.Error()):
-		// Reference not found might be due to an out of sync ledger
-		return true
-	case strings.Contains(msg, orcerrors.ErrNotFound.Error()):
-		// Asset not found might be due to an out of sync ledger
-		return true
-	case strings.Contains(msg, orcerrors.ErrIncompatibleTaskStatus.Error()):
-		// Task status mismatch might be due to an out of sync ledger
 		return true
 	default:
 		return false