+++ server/distributed/server_test.go
@@ -14,9 +14,9 @@ import (
 func TestRetryOnSpecificError(t *testing.T) {
 	assert.False(t, shouldRetry(fmt.Errorf("not an orchestration error")))
 	assert.False(t, shouldRetry(fmt.Errorf("%w", errors.ErrCannotDisableModel)))
-	assert.True(t, shouldRetry(fmt.Errorf("%w", errors.ErrIncompatibleTaskStatus)))
-	assert.True(t, shouldRetry(fmt.Errorf("%w", errors.ErrNotFound)))
-	assert.True(t, shouldRetry(fmt.Errorf("%w", errors.ErrReferenceNotFound)))
+	assert.False(t, shouldRetry(fmt.Errorf("%w", errors.ErrIncompatibleTaskStatus)))
+	assert.False(t, shouldRetry(fmt.Errorf("%w", errors.ErrNotFound)))
+	assert.False(t, shouldRetry(fmt.Errorf("%w", errors.ErrReferenceNotFound)))
 
 	assert.True(t, shouldRetry(fabricTimeout))
 }